+++ CHANGELOG.md
@@ -1,12 +1,29 @@
 # Change Log
 
-## v2.1.0 (xxxx-xx-xx)
+## v2.1.0-beta (2019-12-03)
 
 ### Bug Fixes
 
+无
+
 ### Features
 
++ 新路由中间件取代旧路由中间件（进行中）
+
+#### Next Version
+
++ v2.1.1
+    + auth 适配 new router
+    + rest 适配 new router
+    + ...
+    + router & x/router 使用的 ctx key 更改
++ v2.2.x
+    + 反向代理功能完善
+
 ### Breaking Changes
+
++ router -> x/router
++ xrouter -> router
 
 ## v2.0.0-alpha (xxxx-xx-xx)
 
@@ -14,12 +31,15 @@
 
 + 修复了路由组件匹配上的问题
 + 修复了 auth 中间件 Cookie path 的问题导致的认证失败
++ Option 为 nil 的问题
 
 ### Features
 
 + 合并了 v1 中的一些中间件
 + 新增了中间件： cors、swagger、logger、rest、graphql、(ws)WebSocket、新router
 + 增加独立的 box 仓库存放扩展中间件
++ Context Shadow
++ Use 允许多中间件
 
 ### Breaking Changes